fix: Delete imgae

+++ src/app/API/ToursApiController.js
@@ -1,12 +1,18 @@
 const { Tour } = require("../models/index");
-const fs = require("fs");
+const path = require("path");
+const fs = require("fs-extra");
 
 class TourApiController {
   // [GET] /api/tours
   async findAll(req, res, next) {
     try {
       const tours = await Tour.find({}).lean();
-      res.json({
+      if (!tours)
+        return res
+          .status(404)
+          .json({ success: false, message: "Không tìm thấy tour" });
+
+      res.status(200).json({
         success: true,
         total: tours.length,
         message: "Lấy tours thành công",
@@ -21,7 +27,12 @@ class TourApiController {
   async findTrash(req, res, next) {
     try {
       const tours = await Tour.findWithDeleted({ deleted: true }).lean();
-      res.json({
+      if (!tours)
+        return res
+          .status(404)
+          .json({ success: false, message: "Không tìm thấy tour" });
+
+      res.status(200).json({
         success: true,
         total: tours.length,
         message: "Lấy tours thành công",
@@ -36,7 +47,14 @@ class TourApiController {
   async findOne(req, res, next) {
     try {
       const tour = await Tour.findById(req.params.id).lean();
-      res.json({ success: true, message: "Thành công", data: tour });
+      if (!tour)
+        return res
+          .status(404)
+          .json({ success: false, message: "Không tìm thấy tour" });
+
+      res
+        .status(200)
+        .json({ success: true, message: "Thành công", data: tour });
     } catch (next) {}
   }
 
@@ -61,7 +79,12 @@ class TourApiController {
         thumbnail: imagePaths[0] || "",
         tourType: getTourType(Number(price)),
       });
-      await tour.save();
+      const result = await tour.save();
+      if (!result)
+        return res
+          .status(404)
+          .json({ success: false, message: "Tạo tour thất bại" });
+
       res
         .status(201)
         .json({ success: true, message: "Tạo tour thành công", data: tour });
@@ -73,8 +96,13 @@ class TourApiController {
   // [DELETE] /api/tours/:id
   async softDelete(req, res, next) {
     try {
-      await Tour.delete({ _id: req.params.id });
-      res.json({ success: true, message: "Xoá tour thành công" });
+      const result = await Tour.delete({ _id: req.params.id });
+      if (!result)
+        return res
+          .status(404)
+          .json({ success: false, message: "Xoá tour thất bại" });
+
+      res.status(200).json({ success: true, message: "Xoá tour thành công" });
     } catch (error) {
       next(error);
     }
@@ -83,9 +111,34 @@ class TourApiController {
   // [DELETE] /api/tours/trash/:id
   async delete(req, res, next) {
     try {
-      await Tour.deleteOne({ _id: req.params.id });
-      res.json({ success: true, message: "Xoá tour thành công" });
+      const tour = await Tour.findOneWithDeleted({ _id: req.params.id }).lean();
+
+      if (!tour) {
+        return res.status(404).json({
+          success: false,
+          message: "Không tìm thấy tour",
+        });
+      }
+
+      const result = await Tour.deleteOne({ _id: req.params.id });
+      if (!result)
+        return res
+          .status(404)
+          .json({ success: false, message: "Xoá tour thất bại" });
+
+      if (tour.images && tour.images.length > 0) {
+        try {
+          deleteImages(tour.images);
+        } catch (error) {
+          console.log(error);
+        }
+      }
+      res.status(200).json({
+        success: true,
+        message: "Xoá tour và ảnh thành công",
+      });
     } catch (error) {
+      console.log(error);
       next(error);
     }
   }
@@ -93,12 +146,36 @@ class TourApiController {
   // [PATCH] /api/tours/trash/restore/:id
   async restore(req, res, next) {
     try {
-      await Tour.restore({ _id: req.params.id });
-      res.json({ success: true, message: "Xoá tour thành công" });
+      const result = await Tour.restore({ _id: req.params.id });
+      if (!result)
+        return res
+          .status(404)
+          .json({ success: false, message: "Khôi phục tour thất bại" });
+
+      res
+        .status(200)
+        .json({ success: true, message: "Khôi phục tour thành công" });
     } catch (error) {
       next(error);
     }
   }
+}
+
+// Helpers
+function deleteImages(images) {
+  if (!images || images.length === 0) return;
+  images.forEach((imagePath) => {
+    const fullPath = path.join(__dirname, "../../public", imagePath);
+    if (fs.existsSync(fullPath)) {
+      try {
+        fs.unlinkSync(fullPath);
+      } catch (error) {
+        console.log(error);
+      }
+    } else {
+      console.log(`Không tìm thấy tệp: ${fullPath}`);
+    }
+  });
 }
 
 module.exports = new TourApiController();
